modals

+++ app/ts/directives/explorer/ExplorerFilesController.ts
@@ -1,6 +1,11 @@
+// import * as uib from 'angular-bootstrap';
 import ExplorerFilesScope from './ExplorerFilesScope';
 import IDoodleFile from '../../services/doodles/IDoodleFile';
 import IDoodleManager from '../../services/doodles/IDoodleManager';
+import ModalDialog from '../../services/modalService/ModalDialog';
+import AlertOptions from '../../services/modalService/AlertOptions';
+import ConfirmOptions from '../../services/modalService/ConfirmOptions';
+import PromptOptions from '../../services/modalService/PromptOptions';
 
 /**
  * @class ExplorerFilesController
@@ -12,7 +17,7 @@ export default class ExplorerFilesController {
      * @type string[]
      * @static
      */
-    public static $inject: string[] = ['$scope', 'doodles']
+    public static $inject: string[] = ['$scope', 'doodles', 'modalService']
 
     /**
      * @class ExplorerFilesController
@@ -20,7 +25,7 @@ export default class ExplorerFilesController {
      * @param $scope {ExplorerFilesScope}
      * @param doodles {IDoodleManager}
      */
-    constructor($scope: ExplorerFilesScope, private doodles: IDoodleManager) {
+    constructor($scope: ExplorerFilesScope, private doodles: IDoodleManager, private modalService: ModalDialog) {
         // Define the context menu used by the files.
         $scope.menu = (name: string, file: IDoodleFile) => {
             return [
@@ -37,25 +42,34 @@ export default class ExplorerFilesController {
      * @return {void}
      */
     public newFile(): void {
-        const name = window.prompt("Enter the name of the new file")
-        if (name !== null) {
-            const doodle = this.doodles.current()
-            if (doodle) {
-                try {
-                    doodle.newFile(name)
-                    doodle.selectFile(name)
+        const options: PromptOptions = {
+            title: 'New File',
+            text: '',
+            placeholder: "name.extension",
+            actionButtonText: 'Create File',
+            message: "Enter the name of the new file."
+        };
+        this.modalService.prompt(options)
+            .then((name) => {
+                const doodle = this.doodles.current()
+                if (doodle) {
+                    try {
+                        doodle.newFile(name)
+                        doodle.selectFile(name)
+                        this.doodles.updateStorage()
+                    }
+                    catch (e) {
+                        const alertOptions: AlertOptions = { title: "Error", message: e.toString() }
+                        this.modalService.alert(alertOptions)
+                    }
                 }
-                catch (e) {
-                    window.alert(`${e}`)
+                else {
+                    console.warn(`newFile(${name})`)
                 }
-            }
-            else {
-                console.warn(`newFile(${name})`)
-            }
-        }
-        else {
-            // The user cancelled the request.
-        }
+            })
+            .catch(function(reason: any) {
+                // Do nothing.
+            })
     }
 
     /**
@@ -80,21 +94,32 @@ export default class ExplorerFilesController {
      * @return {void}
      */
     public renameFile(oldName: string): void {
-        const newName = window.prompt("Enter the new name of the file", oldName)
-        if (newName !== null) {
-            const doodle = this.doodles.current()
-            if (doodle) {
-                try {
-                    doodle.renameFile(oldName, newName)
+        const options: PromptOptions = {
+            title: `Rename '${oldName}'`,
+            text: oldName,
+            placeholder: "name.extension",
+            actionButtonText: 'Rename File',
+            message: "Enter the new name of the file."
+        };
+        this.modalService.prompt(options)
+            .then((newName) => {
+                const doodle = this.doodles.current()
+                if (doodle) {
+                    try {
+                        doodle.renameFile(oldName, newName)
+                        this.doodles.updateStorage()
+                    }
+                    catch (e) {
+                        this.modalService.alert({ title: "Error", message: e.toString() })
+                    }
                 }
-                catch (e) {
-                    window.alert(`${e}`)
+                else {
+                    console.warn(`renameFile(${oldName}, ${newName})`)
                 }
-            }
-            else {
-                console.warn(`renameFile(${oldName}, ${newName})`)
-            }
-        }
+            })
+            .catch(function(reason: any) {
+                // Do nothing.
+            })
     }
 
     /**
@@ -103,12 +128,24 @@ export default class ExplorerFilesController {
      * @return {void}
      */
     public deleteFile(name: string): void {
-        const doodle = this.doodles.current()
-        if (doodle) {
-            doodle.deleteFile(name)
-        }
-        else {
-            console.warn(`deleteFile(${name})`)
-        }
+        const options: ConfirmOptions = {
+            title: `Delete '${name}'`,
+            message: `Are you sure you want to delete '${name}'?`,
+            actionButtonText: 'Delete File'
+        };
+        this.modalService.confirm(options)
+            .then((result) => {
+                const doodle = this.doodles.current()
+                if (doodle) {
+                    doodle.deleteFile(name)
+                    this.doodles.updateStorage()
+                }
+                else {
+                    console.warn(`deleteFile(${name})`)
+                }
+            })
+            .catch(function(reason: any) {
+                // Do nothing.
+            })
     }
 }
